replace meaningless graticule tick labels with a .
In the circular world example of #88 the graticule is no longer clipped by the graph box; the original st_graticule tried to do this. It now returns a global graticule, and transforms that; the axis tick values and labels for that case no longer make sense. This commit changes these into "." for this particular case, so they're not longer overplotted by ggplot (they still show as a ".", though).