make st_read|write_db work with non-EPSG srid's
This commit lets st_read_db and st_write_db work with spatial reference systems that are not authorized epsg's, e.g. because the have user-set parameters. In sf, the crs$epsg is dropped (NA) but the proj4string retained. In PostGIS, a new SRID is created beyond the last (usually 900913); the assumption of 900913 being the largest is PostGIS' spatial_ref_sys table is hard coded in sf.