make write_sf overwrite by default, improve docs, #307 #274
modifies the default layer name in case of ESRI Shapefile by stripping a trailing .shp. The GDAL driver does this, but in case we don't do it, delete_layer will not work, as the layer name GDAL writes (without .shp) does not match the one we used to pass to GDAL (with the .shp).