little enhancement of document of st_transform
description of parameter use_gdal: if package `lwgeom` is not installed, the user gets no clue where to find the function `st_transform_proj`. And since there is no hint what this parameter was supposed to do, it's not clear for what reason the user should refer to `st_transform_proj`.  

This enhancement tries to provide clues when to use `st_transform_proj` and where to find it.