Fix issue with key in plot.R
When plotting multiple character/factor-valued columns of an `sf` object, an error occurs when requesting a key because a portion of the code intended for numeric values gets executed (`classInt::classIntervals()`). I added lines to convert character values to a factor, and to check for numeric values before using `classIntervals()`.